refactor(core): move REleasesToolLink out of the ReleasesNav
